% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconrnaseqParam.R
\docType{class}
\name{deconrnaseqParam-class}
\alias{deconrnaseqParam-class}
\title{deconrnaseqParam-class}
\description{
Applies the DeconRNASeq::DeconRNASeq() deconvolution algorithm.
}
\details{
Main constructor for class \linkS4class{deconrnaseqParam}.
}
\examples{
exampleDataList <- getDeconvolutionExampleData()
newParam <- deconrnaseqParam(exampleDataList[["bulkExpression"]],exampleDataList[["referenceExpression"]],exampleDataList[["cellScaleFactors"]])
deconvolution(newParam)

}
\references{
Gong, T., et al. (2011) Optimal Deconvolution of Transcriptional Profiling 
Data Using Quadratic Programming with Application to Complex Clinical Blood 
Samples, PLoS One, 6, e27156.
}
\seealso{
\linkS4class{referencebasedParam-class}
}
