% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconrnaseqParam.R
\name{deconrnaseqParam}
\alias{deconrnaseqParam}
\title{Make new object of class deconrnaseqParam}
\usage{
deconrnaseqParam(
  bulkExpression,
  referenceExpression,
  cellScaleFactors = NULL,
  use.scale = FALSE,
  returnInfo = FALSE
)
}
\arguments{
\item{bulkExpression}{Bulk mixed signals matrix of samples, which can be 
matched to single-cell samples.}

\item{referenceExpression}{Signature matrix of cell type-specific signals. 
If not provided, can be computed from a
provided ExpressionSet containing single-cell data.}

\item{cellScaleFactors}{Cell size factor transformations of length equal to 
the K cell types to deconvolve.}

\item{use.scale}{Whether to scale or center data (default TRUE).}

\item{returnInfo}{Whether to return metadata and original method outputs 
with predicted proportions.}
}
\description{
Main constructor for class \linkS4class{deconrnaseqParam}.
}
\references{
Gong, T., et al. (2011) Optimal Deconvolution of Transcriptional Profiling 
Data Using Quadratic Programming with Application to Complex Clinical Blood 
Samples, PLoS One, 6, e27156.
}
