% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconrnaseqParam.R
\name{deconvolution,deconrnaseqParam-method}
\alias{deconvolution,deconrnaseqParam-method}
\title{Deconvolution method for class \linkS4class{deconrnaseqParam}}
\usage{
\S4method{deconvolution}{deconrnaseqParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{deconrnaseqParam}.}
}
\value{
Either a vector of predicted proportions, or a list containing 
predictions, metadata, and original outputs.
}
\description{
Main deconvolution method for the \linkS4class{deconrnaseqParam} to run the 
\code{DeconRNASeq::DeconRNASeq()} implementation of the DeconRNASeq 
algorithm.
}
\examples{
exampleDataList <- getDeconvolutionExampleData()
newParam <- deconrnaseqParam(exampleDataList[["bulkExpression"]],exampleDataList[["referenceExpression"]],exampleDataList[["cellScaleFactors"]])
deconvolution(newParam)

}
\references{
Gong, T., et al. (2011) Optimal Deconvolution of Transcriptional Profiling 
Data Using Quadratic Programming with Application to Complex Clinical Blood 
Samples, PLoS One, 6, e27156.
}
